Set up build system for class with sample content.
The aim is to have a system not only for generating the class and
documentation, but also a sample document using the class. The system
should demonstrate how to use pandoc with the class, so that the same
Markdown source can be used to generate both a camera-ready PDF and
website-ready HTML.

With this commit, the build system and sample content is in place, but
the class itself is incomplete. (The sample content is actually a real
guide, partially converted from the LaTeX original. In future I expect
to replace it with dummy text.)